+++ src/main/java/nl/nn/testtool/echo2/TestPane.java
@@ -31,16 +31,24 @@ public class TestPane extends Tab implements BeanParent {
 		super();
 	}
 
+	public void setTitle(String title) {
+		this.title = title;
+	}
+
+	public String getTitle() {
+		return title;
+	}
+
 	public void setTreePane(TreePane treePane) {
 		this.treePane = treePane;
 	}
 
-	public void setInfoPane(InfoPane infoPane) {
-		this.infoPane = infoPane;
-	}
-
 	public TreePane getTreePane() {
 		return treePane;
+	}
+
+	public void setInfoPane(InfoPane infoPane) {
+		this.infoPane = infoPane;
 	}
 
 	public InfoPane getInfoPane() {