+++ src/main/java/nl/nn/testtool/web/api/ApiException.java
@@ -24,15 +24,15 @@ import java.lang.invoke.MethodHandles;
 
 public class ApiException extends WebApplicationException implements Serializable {
 	private static final long serialVersionUID = 1L;
-	private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+	private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
 	public ApiException(String msg, Throwable t) {
 		super(t, 500);
-		logger.error(msg, t);
+		log.error(msg, t);
 	}
 
 	public ApiException(String msg) {
 		super(500);
-		logger.error(msg);
+		log.error(msg);
 	}
 }